+++ src/components/Props.jsx
@@ -1,11 +1,11 @@
 import IntroductionToProps from "./props-components/IntroductionToProps";
 import AdditionalPropsTopics from "./props-components/AdditionalPropsTopics";
 
-export default function Props() {
-  function Title() {
-    return <h1>Props</h1>;
-  }
+function Title() {
+  return <h1>Props</h1>;
+}
 
+export default function Props() {
   return (
     <article id='props'>
       <Title />
